requirements.txt
pandas
python-dotenv
requests
streamlit
groq